moving buffer holder per hdmesh
