+++ libs/render_delegate/mesh.cpp
@@ -35,8 +35,6 @@
 #include <pxr/base/gf/vec2f.h>
 #include <pxr/imaging/pxOsd/tokens.h>
 
-#include <mutex>
-
 #include <constant_strings.h>
 #include <shape_utils.h>
 
@@ -174,123 +172,6 @@ inline void _ConvertFaceVaryingPrimvarToBuiltin(
     }
 }
 
-
-
-
-// Compile time mapping of USD type to Arnold types
-template<typename T> inline uint32_t GetArnoldTypeFor(const T &) {return AI_TYPE_UNDEFINED;}
-template<> inline uint32_t GetArnoldTypeFor(const GfVec3f &) {return AI_TYPE_VECTOR;}
-template<> inline uint32_t GetArnoldTypeFor(const VtArray<GfVec3f> &) {return AI_TYPE_VECTOR;}
-template<> inline uint32_t GetArnoldTypeFor(const std::vector<GfVec3f> &) {return AI_TYPE_VECTOR;}
-template<> inline uint32_t GetArnoldTypeFor(const VtArray<int> &) {return AI_TYPE_INT;}
-template<> inline uint32_t GetArnoldTypeFor(const VtArray<unsigned int> &) {return AI_TYPE_UINT;}
-
-
-// A structure to manage and hold the shared buffers. It allows to benefit from USD deduplication passing the same
-// AtArray to arnold when it is presented with the same USD buffer.
-struct BufferHolder {
-    struct HeldArray {
-        HeldArray(uint32_t nref_, const VtValue& val_) : nref(nref_), val(val_) {}
-        uint32_t nref;
-        VtValue val;
-    };
-
-    static std::unordered_map<const void*, HeldArray> _bufferMap;
-    static std::mutex _bufferHolderMutex;
-
-    // TODO we could store the created AtArray and reuse it to benefit from usd deduplication
-    template <typename T>
-    static AtArray* CreateAtArrayFromTimeSamples(const HdArnoldSampledPrimvarType& timeSamples)
-    {
-        if (timeSamples.count == 0)
-            return nullptr;
-
-        // Unbox
-        HdArnoldSampledType<T> unboxed;
-        unboxed.UnboxFrom(timeSamples);
-
-        std::vector<const void*> ptrsToSamples; // use SmallVector ??
-        for (int i = 0; i < unboxed.count; ++i) {
-            const auto& val = unboxed.values[i];
-            ptrsToSamples.push_back(val.data());
-        }
-        const uint32_t nelements = unboxed.values[0].size(); // TODO make sure that values has something
-        const uint32_t type = GetArnoldTypeFor(unboxed.values[0]);
-        const uint32_t nkeys = ptrsToSamples.size();
-        const void** samples = ptrsToSamples.data();
-
-        const std::lock_guard<std::mutex> lock(_bufferHolderMutex);
-
-        AtArray* atArray = AiArrayMakeShared(nelements, nkeys, type, samples, ReleaseArrayCallback, nullptr);
-        if (atArray) {
-            for (int i = 0; i < timeSamples.count; ++i) {
-                const VtValue& val = timeSamples.values[i];
-                if (val.template IsHolding<T>()) {
-                    const T& arr = val.template UncheckedGet<T>();
-                    const void* ptr = static_cast<const void*>(arr.cdata());
-                    auto bufferIt = _bufferMap.find(ptr);
-                    if (bufferIt == _bufferMap.end()) {
-                        _bufferMap.emplace(ptr, HeldArray{1, val});
-                    } else {
-                        HeldArray& held = bufferIt->second;
-                        held.nref++;
-                    }
-                }
-            }
-        }
-        return atArray;
-    }
-
-    template <typename T>
-    static AtArray* CreateAtArrayFromBuffer(
-        const T& vtArray, int32_t forcedType = -1, std::string param = std::string())
-    {
-        const void* arr = static_cast<const void*>(vtArray.cdata());
-        // Look for at AtArray stored with the same buffer
-        if (!arr) {
-            return nullptr;
-        }
-        const uint32_t nelements = vtArray.size();
-        const uint32_t type = forcedType == -1 ? GetArnoldTypeFor(vtArray) : forcedType;
-        
-        const std::lock_guard<std::mutex> lock(_bufferHolderMutex);
-        AtArray*  atArray = AiArrayMakeShared(nelements, type, arr, ReleaseArrayCallback, nullptr);
-        if (atArray) {
-            auto it = _bufferMap.find(arr);
-            if (it == _bufferMap.end()) {
-                _bufferMap.emplace(arr, HeldArray{1, VtValue(vtArray)});
-            } else {
-                // This should rarely happen, only when a single buffer is shared inside keys
-                HeldArray& held = it->second;
-                held.nref++;
-            }
-        }
-        return atArray;
-    }
-
-    static void ReleaseArrayCallback(uint8_t nkeys, const void** buffers, const void* userData)
-    {
-        const std::lock_guard<std::mutex> lock(_bufferHolderMutex);
-        for (int i = 0; i < nkeys; ++i) {
-            const void* arr = buffers[i];
-            if (arr) {
-                auto it = _bufferMap.find(arr);
-                if (it != _bufferMap.end()) {
-                    HeldArray& arr = it->second;
-                    arr.nref--;
-                    if (arr.nref == 0) {
-                        _bufferMap.erase(it);
-                    }
-                } else {
-                    assert(false); // this should never happen, catch it in debug mode
-                }
-            }
-        }
-    }
-};
-std::unordered_map<const void*, BufferHolder::HeldArray> BufferHolder::_bufferMap;
-std::mutex BufferHolder::_bufferHolderMutex;
-
 int HdArnoldSharePositionFromPrimvar(AtNode* node, const SdfPath& id, HdSceneDelegate* sceneDelegate, const AtString& paramName,
     const HdArnoldRenderParam* param, int deformKeys = HD_ARNOLD_DEFAULT_PRIMVAR_SAMPLES,
     const HdArnoldPrimvarMap* primvars = nullptr,  HdArnoldSampledPrimvarType *pointsSample = nullptr,  HdMesh *mesh=nullptr)
@@ -308,8 +189,9 @@ int HdArnoldSharePositionFromPrimvar(AtNode* node, const SdfPath& id, HdSceneDel
         HdArnoldUnboxSample(*pointsSample, xf);
         const auto extrapolatedCount = ExtrapolatePositions(node, paramName, xf, param, deformKeys, primvars);
         if (extrapolatedCount != 0) {
-            // TODO If the points were extrapolated, we used an arnold array and we don't need the pointsSamples anymore,
-            // we need to delete it
+            // If the points were extrapolated, we used an arnold array and we don't need the pointsSamples anymore,
+            // we need to delete its content.
+            pointsSample->Resize(0);
             return extrapolatedCount;
         }
 
@@ -346,9 +228,6 @@ int HdArnoldSharePositionFromPrimvar(AtNode* node, const SdfPath& id, HdSceneDel
                 pointsSample->times[index] = timeSamples[index];
             }
         }
-
-
-        AiNodeSetArray(node, paramName, BufferHolder::CreateAtArrayFromTimeSamples<VtVec3fArray>(*pointsSample));
         return pointsSample->count;
     }
 
@@ -460,7 +339,10 @@ void HdArnoldMesh::Sync(
     } else if (HdChangeTracker::IsPrimvarDirty(*dirtyBits, id, HdTokens->points)) {
         param.Interrupt();
         _numberOfPositionKeys = HdArnoldSharePositionFromPrimvar(node, id, sceneDelegate, str::vlist, param(), GetDeformKeys(), &_primvars, &_pointsSample, this);
-
+        // If the points were extrapolated, _pointsSample is now empty
+        if (_pointsSample.count) {
+            AiNodeSetArray(node, str::vlist, _bufferHolder.CreateAtArrayFromTimeSamples<VtVec3fArray>(_pointsSample));
+        }
     }
 
     bool isLeftHanded = false;
@@ -504,16 +386,16 @@ void HdArnoldMesh::Sync(
             // Keep the buffers alive
             _vertexCountsVtValue = VtValue(vertexCountsTmp);
             _vertexIndicesVtValue = VtValue(vertexIndicesTmp);
-            AiNodeSetArray(GetArnoldNode(), str::nsides, BufferHolder::CreateAtArrayFromBuffer(vertexCountsTmp, AI_TYPE_UINT));
-            AiNodeSetArray(GetArnoldNode(), str::vidxs, BufferHolder::CreateAtArrayFromBuffer(vertexIndicesTmp, AI_TYPE_UINT) );
+            AiNodeSetArray(GetArnoldNode(), str::nsides, _bufferHolder.CreateAtArrayFromBuffer(vertexCountsTmp, AI_TYPE_UINT));
+            AiNodeSetArray(GetArnoldNode(), str::vidxs, _bufferHolder.CreateAtArrayFromBuffer(vertexIndicesTmp, AI_TYPE_UINT) );
 
         } else {
             _vertexCountSum = std::accumulate(vertexCounts.cbegin(), vertexCounts.cend(), 0);
             // Keep the buffers alive
             _vertexCountsVtValue = VtValue(vertexCounts);
             _vertexIndicesVtValue = VtValue(vertexIndices);
-            AiNodeSetArray(GetArnoldNode(), str::nsides, BufferHolder::CreateAtArrayFromBuffer(vertexCounts, AI_TYPE_UINT));
-            AiNodeSetArray(GetArnoldNode(), str::vidxs, BufferHolder::CreateAtArrayFromBuffer(vertexIndices, AI_TYPE_UINT) );
+            AiNodeSetArray(GetArnoldNode(), str::nsides, _bufferHolder.CreateAtArrayFromBuffer(vertexCounts, AI_TYPE_UINT));
+            AiNodeSetArray(GetArnoldNode(), str::vidxs, _bufferHolder.CreateAtArrayFromBuffer(vertexIndices, AI_TYPE_UINT) );
         }
 
         const auto scheme = topology.GetScheme();
